rag_chatbot/requirements.txt
# RAG Chatbot Dependencies

# Core
groq>=0.4.0
streamlit>=1.30.0

# Vector Store
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document Processing
pymupdf>=1.23.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
